Feature/password hash

+++ AI_NETCORE_API/Domain/Creators/SellOffer/Request/Concrete/SellOfferCreateRequest.cs
@@ -7,7 +7,7 @@ namespace Domain.Creators.SellOffer.Request.Concrete
 {
     public class SellOfferCreateRequest : ISellOfferCreateRequest
     {
-        public SellOfferCreateRequest(int resourceId, int amount, decimal price)
+        public SellOfferCreateRequest(int resourceId, int amount, double price)
         {
             ResourceId = resourceId;
             Amount = amount;
@@ -16,7 +16,7 @@ namespace Domain.Creators.SellOffer.Request.Concrete
 
         public int ResourceId { get; }
         public int Amount { get; }
-        public decimal Price { get; }
+        public double Price { get; }
     }
  
 }
